cloud_function_indexer/requirements.txt
functions-framework==3.*
google-cloud-iot
google-cloud-pubsub
google-cloud-bigquery
